Add files via upload
The static approach is utilized as an example of one of these common approaches in the calibration field. The calibration of the ACC is based on the ACC data set gathered at different orientations in static conditions with only gravitational impact. It is advised to collect the outputs of ACC at least in 21 orientations according to the Thin-Shell approach [81] and the reliability of this technique analysis reported in [82]. Moreover, to reduce the effect of temperature drift, it is recommended to warm up the ACC before starting the calibration process, in which it should hold the ACC in each orientation in static conditions and give some time for raw data averaging, which would minimize the impact of random noise. All averages have been conducted along all axes x, y, and z, with at least eight orientations along each axis Figure 3.6, where exact knowledge of these orientations is not necessary [82]. Using the minimization function approach to estimate the nine unknown coefficient of SEM, as a criteria function for minimization, the Root Mean Square Error (RMSE) presented in equation 3.4 has been utilized. This calibration technique assumes that in a static condition where there is no movement the ACC-measured output should equal gravity ‘g’.
Figure 3.6: ACC calibration around eight orientations along “x” axis
In static conditions, the only force impacting the accelerometer is gravitational forces in the downward direction; this fact results in equation 3.3.
g_x^2+g_y^2+g_z^2=|g|^2                                        					         (3.3)
Where g_i   represents the projection of the gravity vector onto axes i. As a criterion function for minimization, RMSE is used. 
 RMSE(Θ,g)=√((∑_1^N▒(|a_i (Θ)|-g)^2 )/N), n=24,                          				        (3.4)
Where   |a_i (Θ)| = √(〖〖acc〗_x〗^2+〖〖acc〗_y〗^2+〖〖acc〗_z〗^2 )        					        (3.5)
and Θ = (Θ_1,Θ_2,…,Θ_9)  is SEM nine coefficient being estimated, the total number of orientations about all axis is N (in our case n=9), m=24 in our case, g=1 is gravity vector, a_i (Θ) represent the SEM acceleration vector being estimated.